# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/remote/songs": {
      "post": {
        "operationId": "SongsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "title": {
                      "type": "string",
                      "example": "string"
                    },
                    "url": {
                      "type": "string",
                      "format": "uri",
                      "example": "https://example.com"
                    },
                    "duration": {
                      "type": "integer",
                      "example": 1
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "PENDING",
                        "APPROVED",
                        "REJECTED"
                      ],
                      "example": "PENDING"
                    }
                  },
                  "required": [
                    "id",
                    "title",
                    "url",
                    "duration",
                    "status"
                  ],
                  "additionalProperties": false,
                  "example": {
                    "id": 1,
                    "title": "string",
                    "url": "https://example.com",
                    "duration": 1,
                    "status": "PENDING"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Songs"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100,
                    "example": "a"
                  },
                  "url": {
                    "type": "string",
                    "format": "uri",
                    "example": "https://example.com"
                  },
                  "duration": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "PENDING",
                      "APPROVED",
                      "REJECTED"
                    ],
                    "example": "PENDING"
                  }
                },
                "required": [
                  "title",
                  "url",
                  "duration"
                ],
                "additionalProperties": false,
                "example": {
                  "title": "a",
                  "url": "https://example.com",
                  "duration": 1
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "SongsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "example": 1
                      },
                      "title": {
                        "type": "string",
                        "example": "string"
                      },
                      "url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://example.com"
                      },
                      "duration": {
                        "type": "integer",
                        "example": 1
                      },
                      "status": {
                        "type": "string",
                        "enum": [
                          "PENDING",
                          "APPROVED",
                          "REJECTED"
                        ],
                        "example": "PENDING"
                      }
                    },
                    "required": [
                      "id",
                      "title",
                      "url",
                      "duration",
                      "status"
                    ],
                    "additionalProperties": false,
                    "example": {
                      "id": 1,
                      "title": "string",
                      "url": "https://example.com",
                      "duration": 1,
                      "status": "PENDING"
                    }
                  },
                  "example": [
                    {
                      "id": 1,
                      "title": "string",
                      "url": "https://example.com",
                      "duration": 1,
                      "status": "PENDING"
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Songs"
        ]
      }
    },
    "/api/remote/songs/{id}": {
      "patch": {
        "operationId": "SongsController_update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "title": {
                      "type": "string",
                      "example": "string"
                    },
                    "url": {
                      "type": "string",
                      "format": "uri",
                      "example": "https://example.com"
                    },
                    "duration": {
                      "type": "integer",
                      "example": 1
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "PENDING",
                        "APPROVED",
                        "REJECTED"
                      ],
                      "example": "PENDING"
                    }
                  },
                  "required": [
                    "id",
                    "title",
                    "url",
                    "duration",
                    "status"
                  ],
                  "additionalProperties": false,
                  "example": {
                    "id": 1,
                    "title": "string",
                    "url": "https://example.com",
                    "duration": 1,
                    "status": "PENDING"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Songs"
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100,
                    "example": "a"
                  },
                  "url": {
                    "type": "string",
                    "format": "uri",
                    "example": "https://example.com"
                  },
                  "duration": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "PENDING",
                      "APPROVED",
                      "REJECTED"
                    ],
                    "example": "PENDING"
                  }
                },
                "additionalProperties": false,
                "example": {
                  "title": "a",
                  "url": "https://example.com",
                  "duration": 1,
                  "status": "PENDING"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "SongsController_remove",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "example": 1
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Songs"
        ]
      }
    },
...
  },
  "info": {
    "title": "PLMA API",
    "description": "PLMA API documentation",
    "version": "0.0",
    "contact": {}
  },
...
  "servers": [],
...
}